refactor: 사용자 활동 outbox 처리 트랜잭션 범위 분리
